Refactor asset history hook and update page
